// src/generated/models/Jia_ll_dshiftgroupsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Jia_ll_dshiftgroupsjia_area = {
  834420000: 'Packing',
  834420001: 'Moulding',
  834420002: 'Processing',
  834420003: 'LFAC'
} as const;
export type Jia_ll_dshiftgroupsjia_area = keyof typeof Jia_ll_dshiftgroupsjia_area;
export const Jia_ll_dshiftgroupsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Jia_ll_dshiftgroupsstatecode = keyof typeof Jia_ll_dshiftgroupsstatecode;
export const Jia_ll_dshiftgroupsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Jia_ll_dshiftgroupsstatuscode = keyof typeof Jia_ll_dshiftgroupsstatuscode;

export interface Jia_ll_dshiftgroupsBase {
  importsequencenumber?: string;
  jia_area?: Jia_ll_dshiftgroupsjia_area;
  jia_department?: string;
  jia_ll_dshiftgroupid: string;
  jia_shift?: string;
  jia_shiftcn?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Jia_ll_dshiftgroupsstatecode;
  statuscode?: Jia_ll_dshiftgroupsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Jia_ll_dshiftgroups extends Jia_ll_dshiftgroupsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  jia_areaname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}